(PC-19593) refactor(navigation): add accessibility tests for PageNotFound
